Refactor classes IbisdocPipe
